LIBFCREPO-811. Require EDTF dates for Letter and Poster content models.
Moved the is_edtf_formatted and is_valid_iso639_code validation functions into the plastron.validation module.

https://issues.umd.edu/browse/LIBFCREPO-811
