+++ plastron/validation/__init__.py
@@ -1,3 +1,20 @@
+from edtf import parse_edtf
+from iso639 import is_valid639_1, is_valid639_2
+from pyparsing import ParseException
+
+
+def is_edtf_formatted(value):
+    try:
+        parse_edtf(value)
+        return True
+    except ParseException:
+        return False
+
+
+def is_valid_iso639_code(value):
+    return is_valid639_1(value) or is_valid639_2(value)
+
+
 class ResourceValidationResult:
     def __init__(self, resource):
         self.resource = resource
